# src/openapi.yaml
openapi: 3.0.1
info:
  title: Explore education statistics - public data API
  version: 1.0.0
  description: >
    This is a draft specification for the upcoming public data API for the EES platform.
  contact:
    name: Explore education statistics
    email: explore.statistics@education.gov.uk
    url: https://explore-education-statistics.service.gov.uk
servers:
  - url: https://ees-api-mock.ambitiousocean-cb084d07.uksouth.azurecontainerapps.io/
    description: Mock API
...
paths:
...
  /api/v1/data-sets/{dataSetId}/query:
    get:
      operationId: QueryDataSetGet
      summary: Query a data set (GET)
      description: |-
        Query a data set using a `GET` request, returning the filtered results.

        Note that there is a `POST` variant of this endpoint which provides a more complete set
        of querying functionality. The `GET` variant is only recommended for initial exploratory 
        testing or simple queries that do not need advanced functionality.

        Unlike the `POST` variant, this endpoint does not allow condition clauses (`and`, `or`, `not`)
        and consequently cannot express more complex queries.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: query
          description: The version of the data set to use e.g. 2.0, 1.1, etc.
          schema:
            type: string
        - name: filters
          in: query
          description: The filters to match results by.
          explode: false
          allowReserved: true
          schema:
            $ref: '#/components/schemas/DataSetQueryCriteriaFilters'
        - name: geographicLevels
          in: query
          description: The geographic levels to match results by.
          explode: false
          allowReserved: true
          schema:
            $ref: '#/components/schemas/DataSetQueryCriteriaGeographicLevels'
        - name: locations
          in: query
          description: The locations to match results by.
          explode: false
          allowReserved: true
          schema:
            $ref: '#/components/schemas/DataSetQueryCriteriaLocations'
        - name: timePeriods
          in: query
          description: The time periods to match results by.
          explode: false
          allowReserved: true
          schema:
            $ref: '#/components/schemas/DataSetQueryCriteriaTimePeriods'
        - name: indicators
          in: query
          description: The IDs or names of indicators to get values for.
          required: true
          explode: false
          allowReserved: true
          schema:
            type: array
            items:
              type: string
        - name: sort
          in: query
          description: |-
            List of sorts to apply to results. Sorts at the start of the list will be applied first.
            By default, results are sorted by time period in descending order.
          explode: false
          allowReserved: true
          schema:
            type: array
            minItems: 1
            items:
              $ref: '#/components/schemas/DataSetQuerySort'
        - name: debug
          in: query
          description: |-
            Enable debug mode. Results will be formatted with human-readable labels to assist identification.
          required: false
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          description: The page of results to fetch.
          required: false
          schema:
            type: number
            format: int32
            minimum: 1
        - name: pageSize
          in: query
          description: The maximum number of results per page.
          required: false
          schema:
            type: number
            format: int32
            minimum: 0
            maximum: 10000
            default: 1000
        - name: Accept
          in: header
          description: |-
            The media type to format the response data in (defaults to `application/json`). Allowed options:

            - `application/json`
            - `text/csv`
          schema:
            type: string
            default: application/json
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetResultsViewModel'
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: QueryDataSetPost
      summary: Query a data set (POST)
      description: |-
        Query a data set using a `POST` request, returning the filtered results.

        Note that for simpler queries or exploratory testing, there is also `GET` variant of this 
        endpoint which only handles a smaller subset of querying functionality. However, for most 
        use-cases, this endpoint is recommended as it provides the complete set of functionality.

        Unlike the `GET` variant, this endpoint allows condition clauses (`and`, `or`, `not`)
        and consequently can express more complex queries.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: query
          description: The version of the data set to use e.g. 2.0, 1.1, etc.
          schema:
            type: string
        - name: debug
          in: query
          description: |-
            Enable debug mode. Results will be formatted with human-readable labels to assist identification.
          required: false
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          description: The page of results to fetch.
          required: false
          schema:
            type: number
            format: int32
            minimum: 1
        - name: pageSize
          in: query
          description: The maximum number of results per page.
          required: false
          schema:
            type: number
            format: int32
            minimum: 0
            maximum: 10000
            default: 1000
        - name: Accept
          in: header
          description: |-
            The media type to format the response data in (defaults to `application/json`). Allowed options:

            - `application/json`
            - `text/csv`
          schema:
            type: string
            default: application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataSetQuery'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetResultsViewModel'
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}/file:
    get:
      operationId: GetDataSetFile
      summary: Get a data set file
      description: Get the entire data set in a specified file format.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: query
          description: The version of the data set to use e.g. 2.0, 1.1, etc.
          schema:
            type: string
        - name: Accept
          in: header
          description: |-
            The media type to output the file in (defaults to `text/csv`). Allowed options:

            - `text/csv` - Outputs the data set as a compressed csv file (defaults to gzip).
            - `application/zip` - Outputs a zip file containing the data set, its metadata and any relevant files.
          schema:
            type: string
            default: text/csv
      responses:
        '200':
          description: Success. A zip containing the underlying data set file.
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/data-sets/{dataSetId}/meta:
    get:
      operationId: GetDataSetMeta
      summary: Get a data set's metadata
      description: Get the metadata about a data set. Use this to create data set queries.
      tags:
        - Data Set
      parameters:
        - name: dataSetId
          in: path
          description: The ID of the data set.
          required: true
          schema:
            type: string
            format: uuid
        - name: dataSetVersion
          in: query
          description: The version of the data set to use e.g. 2.0, 1.1, etc.
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetMetaViewModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  responses:
    BadRequest:
      description: Bad Request. Usually indicates a validation error or a malformed request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorViewModel'
    NotFound:
      description: Not Found. Indicates that a specified resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorViewModel'
  schemas:
    ApiErrorViewModel:
      type: object
      description: Describes an error returned from the API.
      required:
        - title
        - type
        - status
      properties:
        title:
          type: string
          description: The title of the error. Typically summarises the error.
        type:
          type: string
          description: The type of error.
        status:
          type: number
          description: The HTTP status code.
          format: int32
        errors:
          $ref: '#/components/schemas/ErrorDictionary'
...
    DataSetQuery:
      type: object
      description: A query for data from a data set matching specific facets.
      required:
        - facets
        - indicators
      properties:
        facets:
          type: object
          description: A query to match result facets by.
          oneOf:
            - $ref: '#/components/schemas/DataSetQueryCriteria'
            - $ref: '#/components/schemas/DataSetQueryConditionAnd'
            - $ref: '#/components/schemas/DataSetQueryConditionOr'
            - $ref: '#/components/schemas/DataSetQueryConditionNot'
        indicators:
          type: array
          description: The IDs or names of indicators to get values for.
          minItems: 1
          items:
            type: string
        sort:
          type: array
          description: |-
            List of sorts to apply to results. Sorts at the start of the list will be applied first.
            By default, results are sorted by time period in descending order.
          minItems: 1
          items:
            $ref: '#/components/schemas/DataSetQuerySort'
      additionalProperties: false
    DataSetQueryConditionAnd:
      type: object
      description: |-
        A query condition that can only be true when all its sub-clauses resolve to true. This is 
        equivalent to the `AND` operator in SQL.
      required:
        - and
      properties:
        and:
          type: array
          description: The list of sub-clauses.
          minItems: 1
          items:
            oneOf:
              - $ref: '#/components/schemas/DataSetQueryCriteria'
              - $ref: '#/components/schemas/DataSetQueryConditionAnd'
              - $ref: '#/components/schemas/DataSetQueryConditionOr'
              - $ref: '#/components/schemas/DataSetQueryConditionNot'
      additionalProperties: false
      example:
        and:
          - filters:
              in: [filter1, filter2]
          - locations:
              eq: location1
          - timePeriods:
              eq:
                code: AY
                year: 202122
    DataSetQueryConditionOr:
      type: object
      description: |-
        A query condition that can only be true when one of its sub-clauses resolve to true. This
        is equivalent to the `OR` operator in SQL.
      required:
        - or
      properties:
        or:
          type: array
          description: The list of sub-clauses.
          minItems: 1
          items:
            oneOf:
              - $ref: '#/components/schemas/DataSetQueryCriteria'
              - $ref: '#/components/schemas/DataSetQueryConditionAnd'
              - $ref: '#/components/schemas/DataSetQueryConditionOr'
              - $ref: '#/components/schemas/DataSetQueryConditionNot'
      additionalProperties: false
      example:
        or:
          - filters:
              in: [filter1, filter2]
            timePeriods:
              gte:
                code: AY
                year: 201617
              lte:
                code: AY
                year: 201819
          - filters:
              in: [filter1, filter2]
            timePeriods:
              gte:
                code: AY
                year: 202021
              lte:
                code: AY
                year: 202223
    DataSetQueryConditionNot:
      type: object
      description: |-
        A query condition that can only be true when its sub-clause resolves to false. This is 
        equivalent to the `NOT` operator in SQL.
      required:
        - not
      properties:
        not:
          description: The sub-clause that must not be true.
          oneOf:
            - $ref: '#/components/schemas/DataSetQueryCriteria'
            - $ref: '#/components/schemas/DataSetQueryConditionAnd'
            - $ref: '#/components/schemas/DataSetQueryConditionOr'
            - $ref: '#/components/schemas/DataSetQueryConditionNot'
      additionalProperties: false
      example:
        not:
          filters:
            in: [filter1, filter2]
          locations:
            eq: location1
    DataSetQueryCriteria:
      type: object
      description: A set of criteria to match query results to. All parts must resolve to true to match.
      properties:
        filters:
          $ref: '#/components/schemas/DataSetQueryCriteriaFilters'
        locations:
          $ref: '#/components/schemas/DataSetQueryCriteriaLocations'
        geographicLevels:
          $ref: '#/components/schemas/DataSetQueryCriteriaGeographicLevels'
        timePeriods:
          $ref: '#/components/schemas/DataSetQueryCriteriaTimePeriods'
      additionalProperties: false
      example:
        filters:
          in: [filter1, filter2]
        locations:
          notIn: [location1, location2]
        geographicLevels:
          eq: LocalAuthority
        timePeriods:
          gte:
            code: AY
            year: 201617
          lte:
            code: AY
            year: 202021
    DataSetQueryCriteriaFilters:
      type: object
      description: The filter IDs to match results by.
      properties:
        eq:
          type: string
          description: Results must have a filter matching the given ID.
        notEq:
          type: string
          description: Results must not have a filter that matches the given ID.
        in:
          type: array
          description: Results must have at least one filter matching the given IDs.
          minItems: 1
          items:
            type: string
        notIn:
          type: array
          description: Results must not have any filters matching the given IDs.
          minItems: 1
          items:
            type: string
      additionalProperties: false
    DataSetQueryCriteriaTimePeriods:
      type: object
      description: The time periods to match results by.
      properties:
        eq:
          $ref: '#/components/schemas/TimePeriodViewModel'
        notEq:
          $ref: '#/components/schemas/TimePeriodViewModel'
        gte:
          $ref: '#/components/schemas/TimePeriodViewModel'
        gt:
          $ref: '#/components/schemas/TimePeriodViewModel'
        lte:
          $ref: '#/components/schemas/TimePeriodViewModel'
        lt:
          $ref: '#/components/schemas/TimePeriodViewModel'
        in:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/TimePeriodViewModel'
        notIn:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/TimePeriodViewModel'
      additionalProperties: false
    DataSetQueryCriteriaGeographicLevels:
      type: object
      description: The geographic levels to match results by.
      properties:
        eq:
          $ref: '#/components/schemas/GeographicLevel'
        notEq:
          $ref: '#/components/schemas/GeographicLevel'
        in:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/GeographicLevel'
        notIn:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/GeographicLevel'
      additionalProperties: false
    DataSetQueryCriteriaLocations:
      type: object
      description: |-
        The locations to match results by. These can be matched either by the location ID or code. 
        Note the following differences:

        - IDs only match a **single** location
        - Codes may match **multiple** locations

        Whilst codes are mostly unique to a single location, they sometimes are used for multiple 
        locations. This may match more results than you expect so it is recommended to use IDs where
        possible.
      properties:
        eq:
          type: string
          description: Results must have a location matching the given ID or code.
        notEq:
          type: string
          description: Results must not have a location that matches the given ID or code.
        in:
          type: array
          description: Results must have at least one location matching the given IDs or codes.
          minItems: 1
          items:
            type: string
        notIn:
          type: array
          description: Results must not have any locations matching the given IDs or codes.
          minItems: 1
          items:
            type: string
      additionalProperties: false
    DataSetQuerySort:
      type: object
      description: A sorting configuration to sort query results by.
      required:
        - name
      properties:
        name:
          type: string
          description: |-
            The name of the field to sort. This can be `TimePeriod`, a geographic level e.g. `Country` or a filter.
        direction:
          type: string
          description: The direction that the field should be ordered. Defaults to ascending if none is set.
          default: Asc
          enum:
            - Asc
            - Desc
...
    DataSetResultsViewModel:
      type: object
      description: Contains the results of a data set query, and any associated footnotes or meta (where applicable).
      required:
        - footnotes
        - paging
        - results
        - warnings
        - _links
      properties:
        footnotes:
          type: array
          description: The footnotes relevant to the data set query results.
          items:
            $ref: '#/components/schemas/FootnoteViewModel'
        paging:
          $ref: '#/components/schemas/PagingViewModel'
        results:
          type: array
          description: The results of the data set query.
          items:
            $ref: '#/components/schemas/DataSetRowViewModel'
        warnings:
          $ref: '#/components/schemas/WarningDictionary'
        _links:
          $ref: '#/components/schemas/LinksViewModel'
      additionalProperties: false
    DataSetRowViewModel:
      type: object
      description: A row of data point measurements matching a set of facets.
      required:
        - filters
        - geographicLevel
        - locations
        - values
        - timePeriod
      properties:
        id:
          type: string
          description: The ID of the row.
        filters:
          type: object
          description: |-
            A dictionary containing the row's associated filters. 

            The key is the filter name and the value is ID of the filter item.
          additionalProperties:
            type: string
        geographicLevel:
          $ref: '#/components/schemas/GeographicLevel'
        locations:
          type: object
          description: |-
            A dictionary containing the row's associated locations. 

            The key is the geographic level and the value is the ID or code of the location.
          additionalProperties:
            type: string
            description: The ID or code of the location.
        values:
          type: object
          description: |-
            A dictionary containing the row's data values. The key is the indicator name and the value is the measured indicator value.
          additionalProperties:
            type: string
        timePeriod:
          $ref: '#/components/schemas/TimePeriodViewModel'
      additionalProperties: false
    DataSetMetaViewModel:
      type: object
      description: All the metadata associated with a data set.
      required:
        - _links
        - filters
        - indicators
        - geographicLevels
        - locations
        - timePeriods
      properties:
        _links:
          $ref: '#/components/schemas/LinksViewModel'
        filters:
          type: array
          description: All the filters associated with the data set.
          items:
            $ref: '#/components/schemas/FilterMetaViewModel'
        indicators:
          type: array
          description: All the indicators associated with the data set.
          items:
            $ref: '#/components/schemas/IndicatorMetaViewModel'
        geographicLevels:
          type: array
          description: All the geographic levels associated with the data set.
          items:
            $ref: '#/components/schemas/GeographicLevel'
        locations:
          type: array
          description: All the locations associated with the data set, grouped by level.
          items:
            $ref: '#/components/schemas/LocationLevelMetaViewModel'
        timePeriods:
          type: array
          description: All the time periods associated with the data set.
          items:
            $ref: '#/components/schemas/TimePeriodMetaViewModel'
      additionalProperties: false
...
    FilterMetaViewModel:
      type: object
      description: |-
        A group of filterable facets (or characteristics) for the data. This is composed of filter items.
      required:
        - id
        - label
        - options
      properties:
        id:
          type: string
          description: The ID of the filter.
        hint:
          type: string
          description: A hint to assist in describing the filter.
        label:
          type: string
          description: The human-readable label describing the filter.
        options:
          type: array
          description: The filter items belonging to this filter.
          items:
            $ref: '#/components/schemas/FilterOptionMetaViewModel'
      additionalProperties: false
    FootnoteViewModel:
      type: object
      description: A footnote associated with the data.
      required:
        - id
        - content
      properties:
        id:
          type: string
          format: uuid
          description: The ID of the footnote.
        content:
          type: string
          description: The content of the footnote.
      additionalProperties: false
    GeographicLevel:
      type: string
      description: A geographic level that locations are grouped by.
      enum:
        - Country
        - EnglishDevolvedArea
        - LocalAuthority
        - LocalAuthorityDistrict
        - LocalEnterprisePartnership
        - Institution
        - MayoralCombinedAuthority
        - MultiAcademyTrust
        - OpportunityArea
        - ParliamentaryConstituency
        - PlanningArea
        - Provider
        - Region
        - RscRegion
        - School
        - Sponsor
        - Ward
    IndicatorMetaViewModel:
      type: object
      description: The type of measurement taken by a data point.
      required:
        - id
        - label
        - unit
      properties:
        id:
          type: string
          description: The ID of the indicator.
        label:
          type: string
          description: The human-readable label of the indicator.
        unit:
          $ref: '#/components/schemas/Unit'
        decimalPlaces:
          type: integer
          format: int32
          description: The optimal number of decimal places that the indicator should use when displayed.
      additionalProperties: false
    LinkViewModel:
      type: object
      description: A link to the endpoint for a related resource.
      required:
        - href
      properties:
        href:
          type: string
          description: The URI of the link.
        title:
          type: string
          description: The title of the link. Can help understand what the link relates to.
        method:
          type: string
          description: The HTTP method to use with this link.
      example:
        href: 'https://url-to-resource'
    LinksViewModel:
      type: object
      description: |-
        A map of links to related resources/endpoints. A link to the current resource (i.e. `self`) is always provided.
      required:
        - self
      properties:
        self:
          $ref: '#/components/schemas/LinkViewModel'
      additionalProperties:
        $ref: '#/components/schemas/LinkViewModel'
    LocationLevelMetaViewModel:
      type: object
      description: A geographic level's locations.
      required:
        - level
        - options
      properties:
        level:
          $ref: '#/components/schemas/GeographicLevel'
        options:
          type: array
          items:
            $ref: '#/components/schemas/LocationMetaViewModel'
...
    PagingViewModel:
      type: object
      description: Provides metadata for use in pagination.
      additionalProperties: false
      required:
        - page
        - pageSize
        - totalResults
        - totalPages
      properties:
        page:
          type: number
          description: The current page number.
          format: int32
          minimum: 0
        pageSize:
          type: number
          description: The maximum number of results per page.
          format: int32
          minimum: 0
        totalResults:
          type: number
          description: The total number of results across all pages.
          format: int32
          minimum: 0
        totalPages:
          type: number
          description: The total number of pages.
          format: int32
          minimum: 0
...
    TimePeriodCode:
      enum:
        - AY
        - AYQ1
        - AYQ2
        - AYQ3
        - AYQ4
        - CY
        - CYQ1
        - CYQ2
        - CYQ3
        - CYQ4
        - P1
        - P2
        - FY
        - FYQ1
        - FYQ2
        - FYQ3
        - FYQ4
        - TY
        - TYQ1
        - TYQ2
        - TYQ3
        - TYQ4
        - RY
        - T1
        - T1T2
        - T2
        - T3
        - W1
        - W2
        - W3
        - W4
        - W5
        - W6
        - W7
        - W8
        - W9
        - W10
        - W11
        - W12
        - W13
        - W14
        - W15
        - W16
        - W17
        - W18
        - W19
        - W20
        - W21
        - W22
        - W23
        - W24
        - W25
        - W26
        - W27
        - W28
        - W29
        - W30
        - W31
        - W32
        - W33
        - W34
        - W35
        - W36
        - W37
        - W38
        - W39
        - W40
        - W41
        - W42
        - W43
        - W44
        - W45
        - W46
        - W47
        - W48
        - W49
        - W50
        - W51
        - W52
        - M1
        - M2
        - M3
        - M4
        - M5
        - M6
        - M7
        - M8
        - M9
        - M10
        - M11
        - M12
      type: string
      description: |-
        The code describing a time period. This can be one of the following: 

        - `AY` - Academic year
        - `AYQ1-4` - Academic year quarter 1 to 4
        - `CY` - Calendar year
        - `RY` - Reporting year
        - `P1` - Part 1 (April to September)
        - `P2` - Part 2 (October to March)
        - `FY` - Financial year
        - `FYQ1-4` - Financial year quarter 1 to 4
        - `TYQ1-4` - Tax year quarter 1 to 4
        - `W1-52` - Week 1 to 52
        - `M1-12` - Month 1 to 12
...
    TimePeriodViewModel:
      type: object
      description: A compact representation of a time period.
      required:
        - code
        - year
      properties:
        code:
          $ref: '#/components/schemas/TimePeriodCode'
        year:
          type: integer
          format: int32
          description: The year that the time period relates to.
    TimePeriodMetaViewModel:
      type: object
      description: A representation of a time period including a human-readable label.
      allOf:
        - $ref: '#/components/schemas/TimePeriodViewModel'
        - type: object
          required:
            - label
          properties:
            label:
              type: string
              description: The time period in human-readable format.
...
    WarningDictionary:
      type: object
      description: |-
        A dictionary of warnings associated to a model.

        - The key represents a path on the model, or the entire model when it is an empty string (`""`). 
        - The value is a list of warnings associated to the path.
      additionalProperties:
        type: array
        items:
          $ref: '#/components/schemas/WarningViewModel'
...